Move PlaceEncoding to its own file
